requirement.txt
pandas
scikit-learn
nltk
spacy
streamlit
pdfplumber
python-docx
joblib
